fix: 가로 스크롤 탭 오류 수정
